feat: Criando clientes

+++ back/client-service/src/main/java/com/dac/fly/clientservice/repository/TransactionsRepository.java
@@ -9,5 +9,5 @@ import com.dac.fly.clientservice.entity.Transactions;
 
 @Repository
 public interface TransactionsRepository extends JpaRepository<Transactions, Integer> {
-    List<Transactions> findByClienteCodigo(Integer codigoCliente);
+    List<Transactions> findByClienteCodigo(Long codigoCliente);
 }
